+++ app/index.tsx
@@ -1,16 +1,8 @@
 import React, { useEffect } from "react";
-import { Image, StyleSheet, Text, View } from "react-native";
 import { useRouter } from "expo-router";
 import Constants from "expo-constants";
 import Title from "@/components/layouts/Title";
-
-const FlexContainer = ({ children, style }) => {
-  return <View style={[styles.flexContainer, style]}>{children}</View>;
-};
-
-const Logo = () => {
-  return <Image source={require("@/assets/logo.png")} style={styles.logo} />;
-};
+import Logo from "@/components/resources/Logo";
 
 export const StartPage = () => {
   const router = useRouter();
@@ -36,26 +28,5 @@ let AppEntryPoint = StartPage;
 if (Constants.expoConfig?.extra?.storybookEnabled === "true") {
   AppEntryPoint = require("../.ondevice").default;
 }
-
-const styles = StyleSheet.create({
-  flexContainer: {
-    flex: 1,
-    justifyContent: "center",
-    alignItems: "center",
-    backgroundColor: "#0f2422",
-    padding: 16,
-  },
-  logo: {
-    width: 248,
-    height: 248,
-    marginBottom: 16,
-  },
-  title: {
-    fontSize: 48,
-    fontWeight: "bold",
-    color: "#FFFFFF",
-    marginBottom: 32,
-  },
-});
 
 export default AppEntryPoint;
