+++ app/register.tsx
@@ -1,93 +1,15 @@
 import React from "react";
-import { Pressable, ScrollView, StyleSheet, Text, View } from "react-native";
-import { useRouter } from "expo-router";
-import InputLabel from "@/components/Inputs/InputLabel";
-import ButtonSubmit from "@/components/Buttons/ButtonSubmit";
+import { ScrollView, StyleSheet, View } from "react-native";
+import { router, useRouter } from "expo-router";
 import ProgressBarContainer from "@/components/Visuals/ProgressBarContainer";
-import InputDate from "@/components/Inputs/InputDate";
-import RegisterLayout from "@/components/layouts/RegisterLayout";
+import LayoutRegister from "@/components/layouts/register/LayoutRegister";
 import Title from "@/components/layouts/Title";
-
-type RegistrationFormProps = {
-  name: string;
-  setName: (name: string) => void;
-  firstLastName: string;
-  setFirstLastName: (firstLastName: string) => void;
-  secondLastName: string;
-  setSecondLastName: (secondLastName: string) => void;
-  code: string;
-  setCode: (code: string) => void;
-  birthDate: Date | null;
-  setBirthDate: (date: Date | null) => void;
-  onSubmit: () => void;
-};
-
-const FlexContainer: React.FC<{ style?: object }> = ({ children, style }) => {
-  return <View style={[styles.flexContainer, style]}>{children}</View>;
-};
-
-const RegistrationForm: React.FC<RegistrationFormProps> = ({
-  name,
-  setName,
-  firstLastName,
-  setFirstLastName,
-  secondLastName,
-  setSecondLastName,
-  code,
-  setCode,
-  birthDate,
-  setBirthDate,
-  onSubmit,
-}) => {
-  return (
-    <>
-      <InputLabel
-        label="Nombre"
-        value={name}
-        onChangeText={setName}
-        style={styles.userInputMargin}
-      />
-      <InputLabel
-        label="Primer Apellido"
-        value={firstLastName}
-        onChangeText={setFirstLastName}
-        style={styles.userInputMargin}
-      />
-      <InputLabel
-        label="Segundo Apellido"
-        value={secondLastName}
-        onChangeText={setSecondLastName}
-        style={styles.userInputMargin}
-      />
-      <InputLabel
-        label="Código"
-        value={code}
-        onChangeText={setCode}
-        style={styles.userInputMargin}
-      />
-      <InputDate
-        label="Fecha de Nacimiento"
-        value={birthDate}
-        onChange={setBirthDate}
-        style={[styles.userInputMargin, styles.dateInputMargin]} // Ajuste del margen
-      />
-      <ButtonSubmit
-        title="Siguiente"
-        onPress={onSubmit}
-        style={styles.submitButton}
-      />
-    </>
-  );
-};
+import HyperLink from "@/components/Buttons/HyperLink";
+import UserRegistrationForm from "@/components/forms/UserRegistrationForm";
+import HeaderRegister from "@/components/layouts/register/HeaderRegister";
+import ScrollLayout from "@/components/layouts/ScrollLayout";
 
 const Register: React.FC = () => {
-  const [name, setName] = React.useState<string>("");
-  const [firstLastName, setFirstLastName] = React.useState<string>("");
-  const [secondLastName, setSecondLastName] = React.useState<string>("");
-  const [code, setCode] = React.useState<string>("");
-  const [birthDate, setBirthDate] = React.useState<Date | null>(null);
-  const router = useRouter();
-
   const handleSubmit = () => {
     router.push("/register2");
   };
@@ -98,41 +20,27 @@ const Register: React.FC = () => {
 
   return (
     <>
-      <View style={styles.headerContainer}>
+      <HeaderRegister>
         <Title>Registrar</Title>
         <ProgressBarContainer type={"ProgressBar2"} />
-      </View>
-      <RegisterLayout>
-        <ScrollView contentContainerStyle={styles.container}>
-          <RegistrationForm
-            name={name}
-            setName={setName}
-            firstLastName={setFirstLastName}
-            secondLastName={setSecondLastName}
-            setSecondLastName={setSecondLastName}
-            code={code}
-            setCode={setCode}
-            birthDate={birthDate}
-            setBirthDate={setBirthDate}
-            onSubmit={handleSubmit}
+      </HeaderRegister>
+
+      <LayoutRegister>
+        <ScrollLayout>
+          <UserRegistrationForm onSubmit={handleSubmit} />
+          <HyperLink
+            label={"¿Ya tienes cuenta? Inicia Sesión"}
+            onPress={handleLogInRedirect}
           />
-          <Pressable onPress={handleLogInRedirect}>
-            <Text style={styles.logInText}>
-              ¿Ya tienes cuenta? Inicia Sesión
-            </Text>
-          </Pressable>
-        </ScrollView>
-      </RegisterLayout>
+        </ScrollLayout>
+      </LayoutRegister>
     </>
   );
 };
 
 const styles = StyleSheet.create({
-  backgroundContainer: {
-    flex: 1,
-    backgroundColor: "#0f2422",
-  },
   headerContainer: {
+    flex: 1,
     alignItems: "center",
     marginBottom: 20,
     marginTop: 60,
@@ -140,41 +48,6 @@ const styles = StyleSheet.create({
   progressBar: {
     marginTop: 20,
     width: "70%",
-  },
-  formContainerWrapper: {
-    flex: 1,
-    backgroundColor: "#fff",
-    borderTopLeftRadius: 40,
-    paddingTop: 30,
-    paddingBottom: 20,
-    paddingHorizontal: 20,
-    width: "100%",
-  },
-  container: {
-    alignItems: "center",
-  },
-  title: {
-    fontSize: 43.2,
-    fontWeight: "bold",
-    color: "#fff",
-    textAlign: "center",
-  },
-  userInputMargin: {
-    marginTop: 10,
-    width: "100%",
-  },
-  dateInputMargin: {
-    marginTop: 10,
-  },
-  submitButton: {
-    marginTop: 24,
-  },
-  logInText: {
-    color: "#454545",
-    textDecorationLine: "underline",
-    fontSize: 13.3,
-    fontWeight: "bold",
-    marginTop: 24,
   },
 });
