+++ components/Inputs/InputPassword.jsx
@@ -3,16 +3,15 @@ import { Text, TextInput, TouchableOpacity, View } from "react-native";
 import Icon from "react-native-vector-icons/MaterialIcons";
 import styles from "@/styles/Inputs";
 
-const InputPassword = ({ label, value, onChangeText, style }) => {
+const InputPassword = ({ label, onChangeText }) => {
   const [isSecure, setIsSecure] = useState(true);
 
   return (
-    <View style={[styles.container, style]}>
+    <View style={[styles.container]}>
       <Text style={styles.label}>{label}</Text>
       <View style={styles.inputContainer}>
         <TextInput
           style={styles.input}
-          value={value}
           onChangeText={onChangeText}
           secureTextEntry={isSecure}
         />
