+++ components/layouts/Title.tsx
@@ -1,11 +1,12 @@
 import React from "react";
 import { Text } from "react-native";
-import styles from "styles/layouts/Title";
+import styles from "@/styles/layouts/TitleStyle";
 
 type Props = {
   children: string | React.ReactNode;
+  style?: object;
 };
 
-export default function Title({ children }: Props) {
-  return <Text style={styles.title}>{children}</Text>;
+export default function Title({ children, style }: Props) {
+  return <Text style={[styles.title, style]}>{children}</Text>;
 }
