+++ styles/buttons/HyperLinkStyle.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  registerText: {
+    color: "#454545",
+    textDecorationLine: "underline",
+    fontSize: 13.3,
+    fontWeight: "bold",
+    marginTop: 24,
+  },
+});
+
+export default styles;
