+++ styles/forms/LoginForm.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  userInputMargin: {
+    marginTop: 46,
+    marginBottom: 24,
+  },
+  loginButton: {
+    marginTop: 35,
+  },
+});
+
+export default styles;
