+++ styles/forms/RegistrationFormStyle.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  userInputMargin: {
+    marginTop: 10,
+    width: "100%",
+  },
+  dateInputMargin: {
+    marginTop: 10,
+  },
+  submitButton: {
+    marginTop: 24,
+  },
+});
+
+export default styles;
