+++ styles/layouts/ScrollLayoutStyle.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  view: {
+    width: "100%",
+  },
+  container: {
+    alignItems: "center",
+  },
+});
+
+export default styles;
