+++ styles/layouts/register/HeaderRegisterStyle.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  headerContainer: {
+    flex: 1,
+    alignItems: "center",
+    marginBottom: 20,
+    marginTop: 60,
+  },
+});
+
+export default styles;
