+++ styles/layouts/register/RegisterLayoutStyle.js
