+++ styles/resources/Logo.js
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  logo: {
+    width: 248,
+    height: 248,
+    marginBottom: 16,
+  },
+});
+
+export default styles;
